chore: startGameLoop should be triggered by appController after setting up everything
