+++ frontend/tsconfig.json
@@ -4,7 +4,14 @@
     "target": "es2016",
     "moduleResolution": "node",
     "resolveJsonModule": true, // allow importing JSON files
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    // TODO FIX: only work with bundler!!!!
+    // "baseUrl": "./",
+    // "paths": {
+    //   "@/*": [
+    //     "src/*"
+    //   ]
+    // },
   },
   "include": [
     "src/**/*"
